Convert Vec<&GenericSegment> to Vec<Vec<&str>>
Field value extraction is a fairly common task requiring nested
loops with vectors of GenericSegments from searching for all OBX
segments or just pulling the entire generics Vec.

Add a convenience method to Message which takes such a vector as
input and produces a Vec<Vec<&str>> for easy access to &str values.