Implement From<&'a str> for Message
Due to the lifetime tracking required for this library to work as
efficiently as it does, FromStr cannot be implemented as it doesn't
account for the internal str's lifetime. The closest approximation
so far is to implement From<&'a str> returning the Message<'a> to
the caller instead of producing a Result enum like `from_str()`.

Testing
  Added test to compare the structures created by `from(&str)` and
by `from_str(&str)` - passing.