String index - handle MSH off-by-one problem
The HL7 spec defines the first field of the `MSH` segment to be the
field separator directly following the "MSH" characters in the HL7
string. Machines parsing on that delimeter obviously can't handle
such meatland nonsense in format specification natively and need to
have a workaround implemented to offset by -1 for indices into the
`MSH` segment by field.

Implement the off-by-one calculation, and create a temporary hack of
sorts to return `&"|"` because attempts to slice into the `source`
of an MSH segment to pull out `seg[3..3]` violates borrow checker
constraints failing compilation.

@wokket: do you want me to port the change to your `query()` impl?

Relevant spec:
  https://hl7-definition.caristix.com/v2/HL7v2.8/Segments/MSH

Testing:
  Updated MSH string index test to check for MSH.F2 being the set
of delimeters despite that actually being the first field.
  Tests passing