Fix OOB index attempt on fields vec of segment
Under normal `Index<usize>` and friends contexts, indexing out of
bounds is prevented by length checks against the relevant vectors.
However, when extracting a compound index by String or &str, the
decomposition of the compound idx calls into the `fields` Vec when
passing the index operation down to the fields themselves. This
indexing operation does not first check to ensure that it is within
the extents of the vector its touching, resulting in an OOB error
which otherwise appears to be handled in the `usize` versions.

Fix this by adding a bounds check in the compound index destructure
phase and extending tests to cover the condition.

Testing:
  Expanded tests to catch the error
  All tests passing