Refactor `Field` for 3 tiers of subdivision
Thanks @wokket for spec clarification on the internal structures of
a complex HL7 `Field` - the first set of separation occurs in the
repeat delimeter, each of which is then treated effectively as a
whole field. The repeating segments are split like entire `Field`s
were prior to this commit - dividing them up by component delims,
each product of which is split into subcomponent delims.

Functionally, this introduces a new dimension to the data requiring
updates to the `Index` implementations to handle another level of
depth. `Query()` and string indexing are also impacted since their
syntax now maps correctly to the subdivisions of a `Field` struct.

Testing:
  Updated tests for added depth and separation semantic
  All tests passing
  Requires manual review from @wokket since this is a major change
and he knows the spec better.