Expand str idx to take additional level of depth
With the added dimension of repeats pushing existing field vectors
down one level of nesting, the indexing mechanism to quickly pull
values needs to have an additional level of depth.
The Index implementations push down their called parameters to the
lower-level struct within the one handling the current depth of
indexing, so the only effort required to achieve this is to add one
more possible lentgh of index values to the conditional checking
which `uint` indexer to call.
Implement the additional change, add a test to Message extracting
the `S2` subcomponent of `OBR.F1.R1.C2`.

Testing:
  Added test for new functionality
  All tests passing