Implement self-correcting query/index string input
Querying or indexing by string into a Message can produce offset
results due to the mechanism converting the numeric parts of the
query strings into `Index<usize>` calls down the stack. When users
skip elements within the compound query such as "PV1.F1.C1" missing
the "R" element for repeats, the wrong `usize` index is called,
returning the repeat value where the user sought a component value.
This notation produces an index into F1.R1 as the place value which
is supposed to contain F1.R1.C1.

Resolve the concern by having a query string processor construct the
vector of index elements, filling in missing values if and only if
sub-elements are present which would otherwise be offset by one to
the left in the `usize` indexing calls. Add test-case to verify
function.

Testing:
  Test added
  All tests passing